![Update] only use fp16 in sigma/color inference

+++ raymarching/raymarching.py
@@ -21,7 +21,7 @@ from .backend import _backend
 #   rays: int [N, 3], id, offset, num_steps
 class _march_rays_train(Function):
     @staticmethod
-    @custom_fwd(cast_inputs=torch.half)
+    @custom_fwd(cast_inputs=torch.float32)
     def forward(ctx, rays_o, rays_d, bound, density_grid, mean_density, iter_density, step_counter=None, mean_count=-1, perturb=False, align=-1, force_all_rays=False):
         
         rays_o = rays_o.contiguous().view(-1, 3)
@@ -72,7 +72,7 @@ march_rays_train = _march_rays_train.apply
 # outputs: depth: [N], image: [N, 3]
 class _composite_rays_train(Function):
     @staticmethod
-    @custom_fwd(cast_inputs=torch.half)
+    @custom_fwd(cast_inputs=torch.float32)
     def forward(ctx, sigmas, rgbs, deltas, z_vals, rays, bound, bg_color):
         
         sigmas = sigmas.contiguous()
@@ -138,7 +138,7 @@ composite_rays_train = _composite_rays_train.apply
 #   xyzs, dirs, dt: float [n_alive * n_step, 3/3/2], output
 class _march_rays(Function):
     @staticmethod
-    @custom_fwd(cast_inputs=torch.half)
+    @custom_fwd(cast_inputs=torch.float32)
     def forward(ctx, n_alive, n_step, rays_alive, rays_t, rays_o, rays_d, bound, density_grid, mean_density, near, far, align=-1):
         
         rays_o = rays_o.contiguous().view(-1, 3)
@@ -171,7 +171,7 @@ march_rays = _march_rays.apply
 #   depth, image, weight: float [N, 1/3/1]
 class _composite_rays(Function):
     @staticmethod
-    @custom_fwd(cast_inputs=torch.half)
+    @custom_fwd(cast_inputs=torch.float32)
     def forward(ctx, n_alive, n_step, rays_alive, rays_t, sigmas, rgbs, deltas, weights, depth, image):
         _backend.composite_rays(n_alive, n_step, rays_alive, rays_t, sigmas, rgbs, deltas, weights, depth, image)
 
@@ -187,7 +187,7 @@ composite_rays = _composite_rays.apply
 #   rays_t
 class _compact_rays(Function):
     @staticmethod
-    @custom_fwd(cast_inputs=torch.half)
+    @custom_fwd(cast_inputs=torch.float32)
     def forward(ctx, n_alive, rays_alive, rays_alive_old, rays_t, rays_t_old, alive_counter):
         _backend.compact_rays(n_alive, rays_alive, rays_alive_old, rays_t, rays_t_old, alive_counter)
